Move date fixing to utils
