+++ pytroll_collectors/utils.py
@@ -22,6 +22,8 @@
 
 """Utility functions."""
 
+import datetime as dt
+
 from posttroll.publisher import create_publisher_from_dict_config
 
 
@@ -54,3 +56,28 @@ def create_publisher_config_dict(name, nameservers, port):
         'port': port,
         'nameservers': nameservers,
     }
+
+
+def fix_start_end_time(mda):
+    """Make start and end time coherent."""
+    if "duration" in mda and "end_time" not in mda:
+        mda["end_time"] = (mda["start_time"]
+                           + dt.timedelta(seconds=int(mda["duration"])))
+    if "start_date" in mda:
+        mda["start_time"] = dt.datetime.combine(mda["start_date"].date(),
+                                                mda["start_time"].time())
+        if "end_date" not in mda:
+            mda["end_date"] = mda["start_date"]
+        del mda["start_date"]
+    if "end_date" in mda:
+        mda["end_time"] = dt.datetime.combine(mda["end_date"].date(),
+                                              mda["end_time"].time())
+        del mda["end_date"]
+
+    while mda["start_time"] > mda["end_time"]:
+        mda["end_time"] += dt.timedelta(days=1)
+
+    if "duration" in mda:
+        del mda["duration"]
+
+    return mda
